# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_doc_api_v1_contexts__contextId__docs_post": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/DocCategory",
            "default": "internal"
          },
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_doc_api_v1_contexts__contextId__docs_post",
        "type": "object"
      },
...
      "ContextDocResponse": {
        "description": "FE 의 `ContextDoc` 과 같은 모양.\n\n`progress` 는 내보내지 않는다 — 업로드 진행률은 클라이언트만 아는 값이다.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocCategory"
              },
              {
                "type": "null"
              }
            ],
            "description": "기업 컨텍스트 문서의 칸(jd · internal). 이력서는 null."
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/DocKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sizeBytes": {
            "title": "Sizebytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/DocStatus",
            "description": "올라온 직후는 parsing 이다. 본문을 뽑으면 ready, 못 뽑으면 failed. parsing 인 동안 다시 조회한다."
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "sizeBytes",
          "status"
        ],
        "title": "ContextDocResponse",
        "type": "object"
      },
      "ContextResponse": {
        "description": "FE 의 `CompanyContext` 에 `talentProfile` 을 더한 것.\n\nFE 가 「조회가 talentProfile 을 안 돌려줘서 새로 고치면 빈 칸에서 시작한다」고\n한계를 적어 뒀는데(#81), 저장만 되고 못 읽으면 쓸 수 없으므로 여기 싣는다.",
        "properties": {
          "company": {
            "title": "Company",
            "type": "string"
          },
          "docs": {
            "items": {
              "$ref": "#/components/schemas/ContextDocResponse"
            },
            "title": "Docs",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "talentProfile": {
            "title": "Talentprofile",
            "type": "string"
          },
          "team": {
            "title": "Team",
            "type": "string"
          }
        },
        "required": [
          "id",
          "company",
          "team",
          "role",
          "talentProfile",
          "docs"
        ],
        "title": "ContextResponse",
        "type": "object"
      },
...
      "DocCategory": {
        "description": "기업 컨텍스트 문서의 종류. FE 설정 화면의 두 칸이다.\n\nAI 가 JD 와 사내 문서를 다르게 쓴다 — JD 는 「이 직무가 요구하는 것」이다.",
        "enum": [
          "jd",
          "internal"
        ],
        "title": "DocCategory",
        "type": "string"
      },
      "DocKind": {
        "description": "업로드할 수 있는 문서 형식. FE 의 `DocKind` 와 같다.",
        "enum": [
          "pdf",
          "docx"
        ],
        "title": "DocKind",
        "type": "string"
      },
      "DocStatus": {
        "description": "서버가 아는 문서 상태.\n\nFE 의 `DocStatus` 에는 `uploading` 도 있지만 그건 클라이언트에만 있는 상태다 —\n서버는 업로드가 끝난 뒤에야 문서를 안다.\n\n    PARSING --본문을 뽑음--> READY\n            --못 뽑음·한도 초과--> FAILED\n\n올라온 순간 `PARSING` 이고, 백그라운드에서 Helpy 가 본문을 뽑으면 끝난다 (#143).",
        "enum": [
          "parsing",
          "ready",
          "failed"
        ],
        "title": "DocStatus",
        "type": "string"
      },
...
      "ErrorCode": {
        "description": "클라이언트가 분기에 쓰는 값. 추가만 하고 기존 값은 바꾸지 않는다.",
        "enum": [
          "VALIDATION_ERROR",
          "NOT_FOUND",
          "INTERNAL_ERROR",
          "INTERVIEW_NOT_FOUND",
          "SESSION_NOT_FOUND",
          "SESSION_ENDED",
          "INVALID_SESSION_STATE",
          "ROOM_FULL",
          "UNAUTHORIZED",
          "ROLE_NOT_ALLOWED",
          "KAKAO_AUTH_FAILED",
          "KAKAO_UNAVAILABLE"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "default": false,
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "OpenAPI 문서용. 실제 직렬화는 error_body() 가 한다.",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UpdateContextRequest": {
        "description": "`PATCH /contexts/{contextId}` — 넣은 항목만 바꾼다.\n\n길이를 막아 둔다. FE 도 입력창에서 거르지만 그건 편의이지 경계가 아니다 —\n업로드에 같은 원칙을 쓰면서 여기만 열어 두면 앞뒤가 안 맞는다.",
        "properties": {
          "company": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "role": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "talentProfile": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Talentprofile"
          },
          "team": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team"
          }
        },
        "title": "UpdateContextRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "IRYA API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/contexts/current": {
      "get": {
        "description": "지금 쓰는 기업 컨텍스트를 돌려준다. 없으면 만들어서 돌려준다.\n\nFE 가 설정 화면에 들어올 때 `contextId` 를 모르는 상태다. 그래서 id 없이 부를 수\n있는 자리가 하나 필요하다 — FE 도 「조직 컨텍스트를 알려주는 API 가 없어\ncontextId 를 상수로 둔다」고 적어 두고 `'ctx_demo'` 를 쓰고 있다.\n\n**주인은 로그인한 사용자다.** 조직이 아직 없어서 면접관 한 명에 하나다.\n\n없으면 만드는 이유는 「설정을 아직 안 만들었다」와 「설정이 비어 있다」를 FE 가\n구분할 필요가 없어서다. 빈 컨텍스트를 받으면 그냥 채우면 된다.",
        "operationId": "get_current_context_api_v1_contexts_current_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "기업 컨텍스트 조회 (현재 조직)",
        "tags": [
          "기업 컨텍스트"
        ]
      }
    },
    "/api/v1/contexts/{contextId}": {
      "get": {
        "operationId": "get_context_api_v1_contexts__contextId__get",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "title": "Contextid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "컨텍스트를 찾을 수 없음 (남의 컨텍스트 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "기업 컨텍스트 조회",
        "tags": [
          "기업 컨텍스트"
        ]
      },
      "patch": {
        "description": "넣은 항목만 바꾼다. 빈 문자열은 지우라는 뜻이라 그대로 받는다.",
        "operationId": "update_context_api_v1_contexts__contextId__patch",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "title": "Contextid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "컨텍스트를 찾을 수 없음 (남의 컨텍스트 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "요청값 검증 실패"
          }
        },
        "summary": "기업 컨텍스트 수정",
        "tags": [
          "기업 컨텍스트"
        ]
      }
    },
    "/api/v1/contexts/{contextId}/docs": {
      "post": {
        "description": "면접에 쓸 문서를 올린다. **PDF 만 받는다.**\n\nFE 도 보내기 전에 형식과 크기를 거르지만(`UploadRejection`), 그건 편의이지\n경계가 아니다. 여기서 다시 본다.\n\n**응답은 `parsing` 으로 바로 나간다.** 본문 추출(Helpy Document Vision)은 평균\n10초가 걸려서 응답 뒤에 따로 돈다. FE 는 `parsing` 인 문서가 있는 동안\n컨텍스트를 다시 조회하면 `ready` 나 `failed` 로 바뀐 것을 본다.",
        "operationId": "upload_doc_api_v1_contexts__contextId__docs_post",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "title": "Contextid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_doc_api_v1_contexts__contextId__docs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContextDocResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "컨텍스트를 찾을 수 없음 (남의 컨텍스트 포함)"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "파일이 너무 큼"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "지원하지 않는 형식"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 업로드",
        "tags": [
          "기업 컨텍스트"
        ]
      }
    },
    "/api/v1/contexts/{contextId}/docs/{docId}": {
      "delete": {
        "operationId": "delete_doc_api_v1_contexts__contextId__docs__docId__delete",
        "parameters": [
          {
            "in": "path",
            "name": "contextId",
            "required": true,
            "schema": {
              "title": "Contextid",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "docId",
            "required": true,
            "schema": {
              "title": "Docid",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "로그인이 필요함"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "컨텍스트 또는 문서를 찾을 수 없음 (남의 것 포함)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "문서 삭제",
        "tags": [
          "기업 컨텍스트"
        ]
      }
    },
...
  }
}